Plant-Disease-Recognition-System-main/requirements.txt
flask
flask-cors
tensorflow==2.17.0
Pillow
gunicorn